+++ apps/mobile/package.json
@@ -29,8 +29,8 @@
     "@react-native-google-signin/google-signin": "github:react-native-google-signin/google-signin",
     "@react-navigation/bottom-tabs": "^7.0.0",
     "@react-navigation/native": "^7.0.0",
-    "@tamagui/babel-plugin": "^1.121.5",
-    "@tamagui/config": "^1.116.15",
+    "@tamagui/babel-plugin": "1.123.0",
+    "@tamagui/config": "1.123.0",
     "@taquito/utils": "^21.0.0",
     "@umami/core": "workspace:^",
     "@umami/crypto": "workspace:^",
@@ -82,7 +82,7 @@
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "styled-components": "^6.1.13",
-    "tamagui": "^1.120.1",
+    "tamagui": "^1.123.0",
     "util": "^0.12.5",
     "vm-browserify": "^1.1.2"
   },
