+++ apps/mobile/services/auth/Auth.ts
@@ -5,37 +5,33 @@ import { makeRedirectUri } from "expo-auth-session";
 import * as SecureStore from "expo-secure-store";
 import * as WebBrowser from "expo-web-browser";
 
-import { CHAIN_CONFIG_GHOSTNET, WEB3_AUTH_CLIENT_ID } from "../constants";
+import { CHAIN_CONFIG_GHOSTNET, WEB3_AUTH_CLIENT_ID } from "../../constants";
+
+const createWeb3AuthInstance = (): Web3Auth => {
+  const privateKeyProvider = new CommonPrivateKeyProvider({
+    config: { chainConfig: CHAIN_CONFIG_GHOSTNET },
+  });
+
+  const redirectUrl = makeRedirectUri({
+    scheme: "umami",
+    path: "auth",
+  });
+
+  return new Web3Auth(WebBrowser, SecureStore, {
+    clientId: WEB3_AUTH_CLIENT_ID!,
+    network: WEB3AUTH_NETWORK.SAPPHIRE_DEVNET,
+    privateKeyProvider,
+    redirectUrl,
+  });
+};
+
+export const web3auth = createWeb3AuthInstance();
 
 /**
  * Abstract class that's responsible for the social auth process
  * using Web3Auth SDK instead of Torus
  */
 export abstract class Auth {
-  protected web3auth: Web3Auth;
-
-  constructor() {
-    this.web3auth = this.createWeb3AuthInstance();
-  }
-
-  protected createWeb3AuthInstance(): Web3Auth {
-    const privateKeyProvider = new CommonPrivateKeyProvider({
-      config: { chainConfig: CHAIN_CONFIG_GHOSTNET },
-    });
-
-    const redirectUrl = makeRedirectUri({
-      scheme: "umami",
-      path: "auth",
-    });
-
-    return new Web3Auth(WebBrowser, SecureStore, {
-      clientId: WEB3_AUTH_CLIENT_ID!,
-      network: WEB3AUTH_NETWORK.SAPPHIRE_DEVNET,
-      privateKeyProvider,
-      redirectUrl,
-    });
-  }
-
   protected abstract getLoginParams(): LoginParams;
 
   async login(): Promise<{
@@ -47,26 +43,25 @@ export abstract class Auth {
       verifierId?: string;
     };
   }> {
-    if (!this.web3auth.ready) {
-      await this.web3auth.init();
-    }
-
     try {
+      if (!web3auth.connected) {
+        await web3auth.init();
+      }
+
       const loginParams = this.getLoginParams();
-      console.log("loginParams", loginParams);
-      await this.web3auth.login(loginParams);
+      await web3auth.login(loginParams);
     } catch (error) {
       console.error("Error logging in with Web3Auth:", error);
       throw error;
     }
 
-    if (!this.web3auth.connected) {
+    if (!web3auth.connected) {
       throw new Error("Web3Auth connection failed");
     }
 
-    const userInfo = this.web3auth.userInfo();
+    const userInfo = web3auth.userInfo();
 
-    const privateKey = (await this.web3auth.provider?.request({
+    const privateKey = (await web3auth.provider?.request({
       method: "private_key",
     })) as string;
 
@@ -79,6 +74,13 @@ export abstract class Auth {
         verifierId: userInfo?.verifierId || userInfo?.name,
       },
     };
+  }
+
+  async logout() {
+    if (!web3auth.connected) {
+      await web3auth.init();
+    }
+    await web3auth.logout();
   }
 
   async getCredentials(): Promise<{
